dashboard/backend/requirements.txt
flask
flask-cors
pyyaml
mlflow
psycopg2-binary
pandas
requests
gunicorn